Study: 학습 노트 수정 API 및 기능 구현

+++ vue/vue-til/src/components/posts/PostListItem.vue
@@ -1,34 +1,44 @@
 <template>
-  <li >
+  <li>
     <div class="post-title">
-      {{props.postItem.title}}
+      {{ props.postItem.title }}
     </div>
     <div class="post-contents">
-      {{props.postItem.content}}
+      {{ props.postItem.content }}
     </div>
     <div class="post-time">
-      {{props.postItem.createdDate}}
-      <i class="icon ion-md-create"></i>
+      {{ props.postItem.createdDate }}
+      <i class="icon ion-md-create" @click="routeEditPage"></i>
       <i class="icon ion-md-trash" @click="deleteItem"></i>
     </div>
   </li>
 </template>
 
 <script setup>
-  import {deletePost} from "@/api/post";
+import {deletePost} from "@/api/post";
+import {useRouter} from "vue-router";
 
 // eslint-disable-next-line no-undef
 const props = defineProps({
-    postItem:{
-      type:Object,
-      required:true
-    }
-  });
-
-async function deleteItem(){
-    await deletePost(props.postItem.id);
-  console.log('삭제!!')
+  postItem: {
+    type: Object,
+    required: true
   }
+});
+
+// eslint-disable-next-line no-undef
+const emit = defineEmits(['remove']);
+const router = useRouter();
+
+async function deleteItem() {
+  if (!confirm('게시글을 삭제하시겠습니까?')) return;
+  await deletePost(props.postItem.id);
+  emit('remove', props.postItem.id);
+}
+
+function routeEditPage() {
+  router.push(`/post/${props.postItem.id}`);
+}
 
 </script>
